convert 2018-snakemake-cli to python package (setup.py)
converts the simple example in 2018-snakemake-cli,
which used a shell script sitting in a directory to
run the Snakemake workflow, into a full-fledged
python package that uses a setup.py file.

this is a way of bridging the complexity of taco
with the simplicity of 2018-snakemake-cli.

documentation/readme has not yet been updated.

tests have not yet been udpated.